test-requirements.txt
pytest >= 7.2.1
pytest-cov >= 2.8.1
tox >= 3.9.0
flake8 >= 4.0.0
mypy >= 1.5
pyright >= 1.1.0
python-dotenv >= 1.0.0
attrs >= 22.2.0
httpx >= 0.23.0, < 0.29.0
